+++ RomiV_1.2/src/main/java/frc/robot/RobotContainer.java
@@ -1,69 +1,31 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
 package frc.robot;
 
-import edu.wpi.first.wpilibj.GenericHID;
 import edu.wpi.first.wpilibj.XboxController;
 import frc.robot.commands.TankDrive;
-import frc.robot.subsystems.Drivetrain;
+import frc.robot.subsystems.DriveSubsystem;
 import edu.wpi.first.wpilibj2.command.Command;
 import static frc.robot.Constants.*;
 
-
-/**
- * This class is where the bulk of the robot should be declared. Since Command-based is a
- * "declarative" paradigm, very little robot logic should actually be handled in the {@link Robot}
- * periodic methods (other than the scheduler calls). Instead, the structure of the robot (including
- * subsystems, commands, and button mappings) should be declared here.
- */
 public class RobotContainer {
-  
-  // The robot's subsystems and commands are defined here...
-  private final Drivetrain m_drivetrain = new Drivetrain();
 
-  private final XboxController xboxController = new XboxController(IOConstants.XBOX_PORT);
+    private final DriveSubsystem m_drivetrain = new DriveSubsystem();
 
-  // NOTE: The I/O pin functionality of the 5 exposed I/O pins depends on the hardware "overlay"
-  // that is specified when launching the wpilib-ws server on the Romi raspberry pi.
-  // By default, the following are available (listed in order from inside of the board to outside):
-  // - DIO 8 (mapped to Arduino pin 11, closest to the inside of the board)
-  // - Analog In 0 (mapped to Analog Channel 6 / Arduino Pin 4)
-  // - Analog In 1 (mapped to Analog Channel 2 / Arduino Pin 20)
-  // - PWM 2 (mapped to Arduino Pin 21)
-  // - PWM 3 (mapped to Arduino Pin 22)
-  //
-  // Your subsystem configuration should take the overlays into account
+    private final XboxController xboxController = new XboxController(IOConstants.XBOX_PORT);
 
-  /** The container for the robot. Contains subsystems, OI devices, and commands. */
-  public RobotContainer() {
-    m_drivetrain.setDefaultCommand(
-      new TankDrive(
-        m_drivetrain,
-        () -> xboxController.getLeftY(),
-        () -> xboxController.getRightY()
-      )
-    );
+    public RobotContainer() {
+        m_drivetrain.setDefaultCommand(
+        new TankDrive(
+            m_drivetrain,
+            () -> xboxController.getLeftY(),
+            () -> xboxController.getRightY()
+        ));
 
-    // Configure the button bindings
-    configureButtonBindings();
-  }
+        configureButtonBindings();
+    }
 
-  /**
-   * Use this method to define your button->command mappings. Buttons can be created by
-   * instantiating a {@link GenericHID} or one of its subclasses ({@link
-   * edu.wpi.first.wpilibj.Joystick} or {@link XboxController}), and then passing it to a {@link
-   * edu.wpi.first.wpilibj2.command.button.JoystickButton}.
-   */
-  private void configureButtonBindings() {}
+    private void configureButtonBindings() {}
 
-  /**
-   * Use this to pass the autonomous command to the main {@link Robot} class.
-   *
-   * @return the command to run in autonomous
-   */
-  public Command getAutonomousCommand() {
-    return null;
-  }
+    public Command getAutonomousCommand() {
+        return null;
+    }
 }
